requirements.txt
aiohttp==3.8.4
async_timeout==4.0.2
discord.py==2.2.2
Flask==2.2.3
httpx==0.23.3
jishaku==2.5.1
motor==3.1.2
psutil==5.9.4
pymongo==4.3.3
requests==2.26.0
setuptools==67.6.1
tasksio==0.0.0
wavelink==2.1.0
aiosqlite==0.20.0
